refactor: modularize role manager dialog
