+++ components/role-manager/index.ts
@@ -1,0 +1,3 @@
+export { default } from './RoleManagerDialog';
+export { default as RoleManagerDialog } from './RoleManagerDialog';
+export * from './useRoleManager';
